// src/types/ProductTypes/ProductDetails.ts
export type ProductDetails = {
  id: string;
  namespaceId: string;
  category: 'phones' | 'tablets' | 'accessories';
  name: string;
  capacityAvailable: string[];
  capacity: string;
  priceRegular: number;
  priceDiscount: number;
  colorsAvailable: string[];
  color: string;
  images: string[];
  description: {
    title: string;
    text: string[];
  }[];
  screen: string;
  resolution: string;
  processor: string;
  ram: string;
  camera: string;
  zoom: string;
  cell: string[];
};